Fix BUG in mpi_matrix_demo.c
This bug is due to limitations of data generated for matrix a and b.
In original code, we generate a[i][j] = i + j, which means a[i][j]=a[j][i], and same for matrix b.
However, the mistake occurs in MPI matrix computation, in which we are supposed to multiply **a[i][k] to b[k][j]** to get the result. (Priorly written as b[j][k], which is actually same under the test data). 
Now this bug is fixed and we generate the test data RANDOMLY.